Integer addition in python

+++ feature-2.py
@@ -1,5 +1,8 @@
-num1 = float(input("Enter first number: "))
-num2 = float(input("Enter second number: "))
+#person 2 - added this feature change:
+#we have to sum two integer numbers not float 
+
+num1 = int(input("Enter first number: "))
+num2 = int(input("Enter second number: "))
 
 sum = num1 + num2
